lib.nix: Blacklist virto-net module for dpdk guest
This is for consistency with the docker-based environments we have
created before. Can be significant because if the kernel initializes the
Virtio-net device and then the DPDK application takes it over then the
behavior on the Snabb side may be different. (This would be interesting
to have test coverage for too, but at the moment I am more interested in
reproducing our traditional test environment.)